Refactor prebuild command
