+++ packages/expo-cli/src/commands/eject.ts
@@ -6,8 +6,9 @@ import { Command } from 'commander';
 import CommandError from '../CommandError';
 import Log from '../log';
 import { confirmAsync } from '../prompts';
-import * as Eject from './eject/Eject';
+import { ejectAsync } from './eject/ejectAsync';
 import { platformsFromPlatform } from './eject/platformOptions';
+import { EjectAsyncOptions } from './eject/prebuildAsync';
 import { learnMore } from './utils/TerminalLink';
 
 async function userWantsToEjectWithoutUpgradingAsync() {
@@ -23,7 +24,7 @@ export async function actionAsync(
   {
     platform,
     ...options
-  }: Omit<Eject.EjectAsyncOptions, 'platforms'> & {
+  }: Omit<EjectAsyncOptions, 'platforms'> & {
     npm?: boolean;
     platform?: string;
   }
@@ -43,10 +44,10 @@ export async function actionAsync(
     }
   } else {
     Log.debug('Eject Mode: Latest');
-    await Eject.ejectAsync(projectDir, {
+    await ejectAsync(projectDir, {
       ...options,
       platforms: platformsFromPlatform(platform),
-    } as Eject.EjectAsyncOptions);
+    } as EjectAsyncOptions);
   }
 }
