+++ packages/expo-cli/src/commands/eject/__tests__/createNativeProjectsFromTemplateAsync-test.ts
@@ -1,54 +1,20 @@
 import { vol } from 'memfs';
 
 import {
-  ensureValidPlatforms,
   hashForDependencyMap,
   isPkgMainExpoAppEntry,
   resolveBareEntryFile,
   shouldDeleteMainField,
-  stripDashes,
-} from '../Eject';
+} from '../createNativeProjectsFromTemplateAsync';
 
 jest.mock('fs');
 jest.mock('resolve-from');
-
-describe('stripDashes', () => {
-  it(`removes spaces and dashes from a string`, () => {
-    expect(stripDashes(' My cool-app ')).toBe('Mycoolapp');
-    expect(stripDashes(' --- - ----- ')).toBe('');
-    expect(stripDashes('-----')).toBe('');
-    expect(stripDashes(' ')).toBe('');
-    expect(stripDashes(' \n-\n-')).toBe('');
-  });
-});
 
 describe('hashForDependencyMap', () => {
   it(`dependencies in any order hash to the same value`, () => {
     expect(hashForDependencyMap({ a: '1.0.0', b: 2, c: '~3.0' })).toBe(
       hashForDependencyMap({ c: '~3.0', b: 2, a: '1.0.0' })
     );
-  });
-});
-
-describe(ensureValidPlatforms, () => {
-  const platform = process.platform;
-
-  afterEach(() => {
-    Object.defineProperty(process, 'platform', {
-      value: platform,
-    });
-  });
-  it(`bails on windows if only ios is passed`, async () => {
-    Object.defineProperty(process, 'platform', {
-      value: 'win32',
-    });
-    expect(ensureValidPlatforms(['ios', 'android'])).toStrictEqual(['android']);
-  });
-  it(`allows ios on all platforms except windows`, async () => {
-    Object.defineProperty(process, 'platform', {
-      value: 'other',
-    });
-    expect(ensureValidPlatforms(['ios', 'android'])).toStrictEqual(['ios', 'android']);
   });
 });
