+++ packages/expo-cli/src/commands/eject/platformOptions.ts
@@ -1,4 +1,5 @@
 import { ModPlatform } from '@expo/config-plugins';
+import chalk from 'chalk';
 
 import CommandError from '../../CommandError';
 import Log from '../../log';
@@ -28,5 +29,26 @@ export function platformsFromPlatform(platform?: string): ModPlatform[] {
       return getDefaultPlatforms();
     default:
       throw new CommandError(`Unsupported platform "${platform}". Options are: ios, android, all`);
+  }
+}
+
+export function ensureValidPlatforms(platforms: ModPlatform[]): ModPlatform[] {
+  const isWindows = process.platform === 'win32';
+  // Skip ejecting for iOS on Windows
+  if (isWindows && platforms.includes('ios')) {
+    Log.warn(
+      `⚠️  Skipping generating the iOS native project files. Run ${chalk.bold(
+        'expo eject'
+      )} again from macOS or Linux to generate the iOS project.`
+    );
+    Log.newLine();
+    return platforms.filter(platform => platform !== 'ios');
+  }
+  return platforms;
+}
+
+export function assertPlatforms(platforms: ModPlatform[]) {
+  if (!platforms?.length) {
+    throw new CommandError('At least one platform must be enabled when syncing');
   }
 }
